Add comprehensive tests for versioning CLI and publication features
- Implement CLI tests for version bump, sync, release commands, and argument parsing in `test_cli.py`.
- Create tests for HAL publication service and CLI publication commands in `test_publications.py`.
- Add release service tests covering dry runs, metadata syncing, and contributor handling in `test_release.py`.
- Introduce repository tests for version state reading, bumping, and changelog synchronization in `test_repository.py`.

feat(ops): add HAL publication harvesting to fpp-version release notes
Fixes #2454
